refactor(goods): 도메인 의존성 정리를 위해 SelectedOption 타입을 entities/goods 로 이동
