--- src/shared/types/selectedOption.ts
@@ -1,6 +1,0 @@
-import { GoodDetailOption } from "@/entities/goods/goods.types";
-
-export type SelectedOption = {
-  type: "option" | "extraOption";
-  quantity: number;
-} & GoodDetailOption;
